remove superfulous spaces, use uintmax for fq-rate

+++ include/util.h
@@ -65,15 +65,15 @@ extern "C" {
  * set/getsockopt wrappers for SO_RCVBUF and SO_SNDBUF; TCP_MAXSEG
  * socket.c
  * ------------------------------------------------------------------- */
-int setsock_tcp_windowsize( int inSock, int inTCPWin, int inSend );
-int getsock_tcp_windowsize( int inSock, int inSend );
+int setsock_tcp_windowsize(int inSock, int inTCPWin, int inSend);
+int getsock_tcp_windowsize(int inSock, int inSend);
 
-void setsock_tcp_mss( int inSock, int inTCPWin );
-int  getsock_tcp_mss( int inSock );
+void setsock_tcp_mss(int inSock, int inTCPWin);
+int  getsock_tcp_mss(int inSock);
 bool setsock_blocking(int fd, bool blocking);
 
-int recvn( int inSock, char *outBuf, int inLen, int flags );
-int writen( int inSock, const void *inBuf, int inLen);
+int recvn(int inSock, char *outBuf, int inLen, int flags);
+int writen(int inSock, const void *inBuf, int inLen);
 
 void disarm_itimer(void);
 /* -------------------------------------------------------------------
@@ -81,11 +81,11 @@ void disarm_itimer(void);
  * signal.c
  * ------------------------------------------------------------------- */
 typedef void Sigfunc(int);
-void sig_exit( int inSigno );
+void sig_exit(int inSigno);
 
 typedef Sigfunc *SigfuncPtr;
 
-SigfuncPtr my_signal( int inSigno, SigfuncPtr inFunc );
+SigfuncPtr my_signal(int inSigno, SigfuncPtr inFunc);
 
 #ifdef WIN32
 
@@ -101,7 +101,7 @@ enum {
 };
 #endif
 
-BOOL WINAPI sig_dispatcher( DWORD type );
+BOOL WINAPI sig_dispatcher(DWORD type);
 
 #endif
 
@@ -109,72 +109,72 @@ BOOL WINAPI sig_dispatcher( DWORD type );
  * error handlers
  * error.c
  * ------------------------------------------------------------------- */
-void warn      ( const char *inMessage, const char *inFile, int inLine );
-void warn_errno( const char *inMessage, const char *inFile, int inLine );
+void warn      (const char *inMessage, const char *inFile, int inLine);
+void warn_errno(const char *inMessage, const char *inFile, int inLine);
 
-#if defined( HAVE_POSIX_THREAD ) || defined( HAVE_WIN32_THREAD)
-#define FAIL( cond, msg, settings )             \
+#if defined(HAVE_POSIX_THREAD) || defined(HAVE_WIN32_THREAD)
+#define FAIL(cond, msg, settings)             \
   do {                                          \
-    if ( cond ) {                               \
-      warn( msg, __FILE__, __LINE__ );          \
+    if (cond) {                               \
+      warn(msg, __FILE__, __LINE__);          \
       thread_stop(settings);                    \
     }                                           \
-  } while( 0 )
+  } while(0)
 #else
-#define FAIL( cond, msg, settings )             \
+#define FAIL(cond, msg, settings)             \
   do {                                          \
-    if ( cond ) {                               \
-      warn( msg, __FILE__, __LINE__ );          \
-      exit( 1 );                                \
+    if (cond) {                               \
+      warn(msg, __FILE__, __LINE__);          \
+      exit(1);                                \
     }                                           \
-  } while( 0 )
+  } while(0)
 #endif
 
-#define WARN( cond, msg )                       \
+#define WARN(cond, msg)                       \
   do {                                          \
-    if ( cond ) {                               \
-      warn( msg, __FILE__, __LINE__ );          \
+    if (cond) {                               \
+      warn(msg, __FILE__, __LINE__);          \
     }                                           \
-  } while( 0 )
+  } while(0)
 
-#if defined( HAVE_POSIX_THREAD ) || defined( HAVE_WIN32_THREAD)
-#define FAIL_errno( cond, msg, settings )       \
+#if defined(HAVE_POSIX_THREAD) || defined(HAVE_WIN32_THREAD)
+#define FAIL_errno(cond, msg, settings)       \
   do {                                          \
-    if ( cond ) {                               \
-      warn_errno( msg, __FILE__, __LINE__ );    \
+    if (cond) {                               \
+      warn_errno(msg, __FILE__, __LINE__);    \
       thread_stop(settings);                    \
     }                                           \
-  } while( 0 )
+  } while(0)
 #else
-#define FAIL_errno( cond, msg, settings )       \
+#define FAIL_errno(cond, msg, settings)       \
   do {                                          \
-    if ( cond ) {                               \
-      warn_errno( msg, __FILE__, __LINE__ );    \
-      exit( 1 );                                \
+    if (cond) {                               \
+      warn_errno(msg, __FILE__, __LINE__);    \
+      exit(1);                                \
     }                                           \
-  } while( 0 )
+  } while(0)
 #endif
 
-#define WARN_errno( cond, msg )                 \
+#define WARN_errno(cond, msg)                 \
   do {                                          \
-    if ( cond ) {                               \
-      warn_errno( msg, __FILE__, __LINE__ );    \
+    if (cond) {                               \
+      warn_errno(msg, __FILE__, __LINE__);    \
     }                                           \
-  } while( 0 )
+  } while(0)
 
 /* -------------------------------------------------------------------
  * initialize buffer to a pattern
  * ------------------------------------------------------------------- */
-void pattern( char *outBuf, int inBytes );
+void pattern(char *outBuf, int inBytes);
 
 /* -------------------------------------------------------------------
  * input and output numbers, converting with kilo, mega, giga
  * stdio.c
  * ------------------------------------------------------------------- */
-double byte_atof( const char *inString );
-double bitorbyte_atof( const char *inString );
-intmax_t byte_atoi( const char  *inString );
-unsigned int bitorbyte_atoi( const char *inString );
+double byte_atof(const char *inString);
+double bitorbyte_atof(const char *inString);
+intmax_t byte_atoi(const char  *inString);
+uintmax_t bitorbyte_atoi(const char *inString);
 void byte_snprintf(char* outString, int inLen, double inNum, char inFormat);
 
 /*
@@ -184,19 +184,19 @@ void byte_snprintf(char* outString, int inLen, double inNum, char inFormat);
 
 #define TimeZero(timeval) ((timeval.tv_sec == 0) && (timeval.tv_usec == 0))
 
-#define TimeDifference( left, right ) (left.tv_sec  - right.tv_sec) +   \
+#define TimeDifference(left, right) (left.tv_sec  - right.tv_sec) +   \
         (left.tv_usec - right.tv_usec) / ((double) rMillion)
 
 #define TimeDouble(timeval) (timeval.tv_sec + timeval.tv_usec / ((double) rMillion))
 
-#define TimeAdd( left, right )  do {                                    \
+#define TimeAdd(left, right)  do {                                    \
                                     left.tv_usec += right.tv_usec;      \
-                                    if ( left.tv_usec > rMillion ) {    \
+                                    if (left.tv_usec > rMillion) {    \
                                         left.tv_usec -= rMillion;       \
                                         left.tv_sec++;                  \
                                     }                                   \
                                     left.tv_sec += right.tv_sec;        \
-                                } while ( 0 )
+                                } while (0)
 
 /* -------------------------------------------------------------------
  * redirect the stdout to a specified file
@@ -207,29 +207,29 @@ void redirect(const char *inOutputFileName);
 /* -------------------------------------------------------------------
  * delete macro
  * ------------------------------------------------------------------- */
-#define DELETE_PTR( ptr )                       \
+#define DELETE_PTR(ptr)                       \
   do {                                          \
-    if ( ptr != NULL ) {                        \
+    if (ptr != NULL) {                        \
       delete ptr;                               \
       ptr = NULL;                               \
     }                                           \
-  } while( false )
+  } while(false)
 
-#define DELETE_ARRAY( ptr )                     \
+#define DELETE_ARRAY(ptr)                     \
   do {                                          \
-    if ( ptr != NULL ) {                        \
+    if (ptr != NULL) {                        \
       delete [] ptr;                            \
       ptr = NULL;                               \
     }                                           \
-  } while( false )
+  } while(false)
 
-#define FREE_ARRAY( ptr )                     \
+#define FREE_ARRAY(ptr)                     \
   do {                                          \
-    if ( ptr != NULL ) {                        \
+    if (ptr != NULL) {                        \
       free(ptr); \
       ptr = NULL;                               \
     }                                           \
-  } while( false )
+  } while(false)
 
 #ifdef __cplusplus
 } /* end extern "C" */
